meaning of laser 'power' is always average power
Previously the 'power' attribute of either laser (e.g., as passed in through
the yaml parameter file) was always treated as the power whenever the laser was
on (e.g., peak power for a pulsed laser). It's more natural to always pass in
the average power, for both pulsed and CW lasers, and then calculate the peak
power for pulsed lasers as needed. Changes to implement this:

- create peakpower function in ohcalcs
- create basic unit tests for peakpower in test_ohcalcs
- call peakpower from KineticsRun.intensity to calculate a peakpower to pass to
spec_intensity if the laser is pulsed.
- add a reprate field to the yaml template for both irlaser and uvlaser that
KineticsRun.intensity passes to ohcalcs.peakpower if the laser is pulsed.